move edit.js into edit folder
