+++ src/superlist/index.js
@@ -19,7 +19,7 @@ import "./style.scss";
  * Internal dependencies
  */
 import metadata from "./block.json";
-import edit from "./edit";
+import edit from "./edit/edit";
 import save from "./save";
 import example from "./example/example";
 import transforms from "./transforms";
